Final updates
Edits to the copy in the index.html file as well as a load of updates to the CSS to make everything high and tight! Looking good!

Only thing left is to decide the fate of that "connect with me" section at the bottom - I've gone the route of the icons, but haven't linked them yet.